+++ Stripe/STPPaymentMethodCardParams.m
@@ -1,0 +1,31 @@
+//
+//  STPPaymentMethodCardParams.m
+//  Stripe
+//
+//  Created by Yuki Tokuhiro on 3/6/19.
+//  Copyright © 2019 Stripe, Inc. All rights reserved.
+//
+
+#import "STPPaymentMethodCardParams.h"
+
+@implementation STPPaymentMethodCardParams
+
+@synthesize additionalAPIParameters = _additionalAPIParameters;
+
+#pragma mark - STPFormEncodable
+
++ (NSString *)rootObjectName {
+    return @"card";
+}
+
++ (NSDictionary *)propertyNamesToFormFieldNamesMapping {
+    return @{
+             NSStringFromSelector(@selector(number)): @"number",
+             NSStringFromSelector(@selector(expMonth)): @"exp_month",
+             NSStringFromSelector(@selector(expYear)): @"exp_year",
+             NSStringFromSelector(@selector(cvc)): @"cvc",
+             NSStringFromSelector(@selector(token)): @"token",
+             };
+}
+
+@end