+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   * Renames `STPPaymentMethodTuple` to `STPPaymentOptionTuple`
   * Renames `STPPaymentMethodsViewController` to `STPPaymentOptionsViewController`
   * Renames all properties, methods, comments referencing 'PaymentMethod' to 'PaymentOption'
+* Adds `[STPAPI createPaymentMethodWithParams:completion:]`, which creates a PaymentMethod.
 
 ## 14.0.0 2018-11-14
 * Changes `STPPaymentCardTextField`, which now copies the `cardParams` property. See [MIGRATING.md](/MIGRATING.md) for more details. [#1031](https://github.com/stripe/stripe-ios/pull/1031)